Refactor: extract the Merging of Agents into an ExecutionPlanOptimizer abstraction
