+++ api/src/main/java/com/datastax/oss/sga/api/runtime/AgentNodeProvider.java
@@ -32,15 +32,4 @@ public interface AgentNodeProvider {
      */
     boolean supports(String type, ComputeClusterRuntime clusterRuntime);
 
-    /**
-     * Returns the ability of an Agent to be merged with the previous version.
-     * @return true if the agents can be merged
-     */
-    default boolean canMerge(AgentNode previousAgent, AgentNode agentImplementation) {
-        return false;
-    }
-
-    default AgentNode mergeAgents(AgentNode previousAgent, AgentNode agentImplementation, ExecutionPlan instance) {
-        throw new UnsupportedOperationException("This provider cannot merge agents");
-    }
 }
