+++ core/src/main/java/com/datastax/oss/sga/impl/agents/AbstractCompositeAgentProvider.java
@@ -1,4 +1,4 @@
-package com.datastax.oss.sga.runtime.impl.k8s.agents;
+package com.datastax.oss.sga.impl.agents;
 
 import com.datastax.oss.sga.api.model.AgentConfiguration;
 import com.datastax.oss.sga.api.runtime.ComponentType;
@@ -8,8 +8,6 @@ import lombok.extern.slf4j.Slf4j;
 import java.util.List;
 import java.util.Set;
 
-import static com.datastax.oss.sga.runtime.impl.k8s.KubernetesClusterRuntime.CLUSTER_TYPE;
-
 /**
  * Implements support for processors that can be executed in memory into a single pipeline.
  * This is a special processor that executes a pipeline of Agents in memory.
@@ -17,14 +15,14 @@ import static com.datastax.oss.sga.runtime.impl.k8s.KubernetesClusterRuntime.CLU
  * It is created internally by the planner.
  */
 @Slf4j
-public class CompositeAgentProvider extends AbstractAgentProvider {
+public abstract class AbstractCompositeAgentProvider extends AbstractAgentProvider {
 
     public static final String AGENT_TYPE = "composite-agent";
 
     private static final Set<String> SUPPORTED_AGENT_TYPES = Set.of(AGENT_TYPE);
 
-    public CompositeAgentProvider() {
-        super(SUPPORTED_AGENT_TYPES, List.of(CLUSTER_TYPE, "none"));
+    public AbstractCompositeAgentProvider(List<String> clusterRuntimes) {
+        super(SUPPORTED_AGENT_TYPES, clusterRuntimes);
     }
 
     @Override
