+++ k8s-runtime/k8s-runtime-core/src/main/java/com/datastax/oss/sga/runtime/impl/k8s/agents/TextProcessingAgentsProvider.java
@@ -2,6 +2,7 @@ package com.datastax.oss.sga.runtime.impl.k8s.agents;
 
 import com.datastax.oss.sga.api.model.AgentConfiguration;
 import com.datastax.oss.sga.api.runtime.ComponentType;
+import com.datastax.oss.sga.impl.agents.AbstractComposableAgentProvider;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.List;
