+++ k8s-runtime/k8s-runtime-core/src/main/java/com/datastax/oss/sga/runtime/impl/k8s/agents/ai/KubernetesGenAIToolKitFunctionAgentProvider.java
@@ -6,7 +6,7 @@ import com.datastax.oss.sga.runtime.impl.k8s.KubernetesClusterRuntime;
 public class KubernetesGenAIToolKitFunctionAgentProvider extends GenAIToolKitFunctionAgentProvider {
 
     public KubernetesGenAIToolKitFunctionAgentProvider() {
-        super(KubernetesClusterRuntime.CLUSTER_TYPE, "ai-tools");
+        super(KubernetesClusterRuntime.CLUSTER_TYPE, GenAIToolKitFunctionAgentProvider.AGENT_TYPE);
     }
 
 }
