+++ k8s-runtime/k8s-runtime-core/src/main/resources/META-INF/services/com.datastax.oss.sga.api.runtime.AgentNodeProvider
@@ -2,4 +2,4 @@ com.datastax.oss.sga.runtime.impl.k8s.agents.python.PythonCodeAgentProvider
 com.datastax.oss.sga.runtime.impl.k8s.agents.ai.KubernetesGenAIToolKitFunctionAgentProvider
 com.datastax.oss.sga.runtime.impl.k8s.agents.S3SourceAgentProvider
 com.datastax.oss.sga.runtime.impl.k8s.agents.TextProcessingAgentsProvider
-com.datastax.oss.sga.runtime.impl.k8s.agents.CompositeAgentProvider
+com.datastax.oss.sga.runtime.impl.k8s.agents.KubernetesCompositeAgentProvider
